+++ public/src/scripts/html-content/index.js
@@ -1,5 +1,5 @@
 import enterForm from "./enter-form";
-import mockGame from "./maze-matrix";
+import mazeMatrix from "./maze-matrix";
 import { maze } from "./maze";
 
-export { enterForm, mockGame, maze };
+export { enterForm, mazeMatrix, maze };
